requirements.txt
streamlit
pandas
torch
numpy
scikit-learn
matplotlib
plotly
fpdf
openpyxl  # necesario para leer archivos Excel (.xlsx)
Pillow     # para trabajar con imágenes